Add emitter contract.
