+++ crates/action-runner/src/events.rs
@@ -1,0 +1,9 @@
+pub enum Event {
+    TargetOutputIsCached,
+}
+
+struct Emitter {}
+
+impl Emitter {
+    pub fn test() {}
+}
